Include some of the payload installation tasks by default
Some of the installation tasks don't depend on the type of the payload
module, so include them by default on the service-level.
