+++ tests/nosetests/pyanaconda_tests/modules/payloads/payload/module_payload_live_image_test.py
@@ -149,7 +149,6 @@ class LiveImageModuleTestCase(unittest.TestCase):
     def post_install_with_tasks_test(self, publisher):
         """Test Live Image post installation configuration task."""
         # task_classes = [
-        #     CopyDriverDisksFilesTask,
         #     TeardownInstallationSourceImageTask
         # ]
         #
